+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## 1.0.0-beta8
 
 - Removed Rasterio dependency, instead using GDAL directly
+- Pass FeatureClass and FeatureDataset sequences on instantiation
 - Getter methods return dicts instead of tuples
 - GDAL raster support by default in conda, optional in pip install
 - Dev environment setup/test/build batch scripts